Move config functions to config package
