+++ config/appconfig_test.go
@@ -1,8 +1,9 @@
-package goa4web
+package config_test
 
 import (
 	"testing"
 
+	"github.com/arran4/goa4web/config"
 	"github.com/arran4/goa4web/core"
 )
 
@@ -13,7 +14,7 @@ func TestLoadAppConfigFile(t *testing.T) {
 	if err := fs.WriteFile(file, []byte(content), 0644); err != nil {
 		t.Fatalf("write config: %v", err)
 	}
-	m := LoadAppConfigFile(fs, file)
+	m := config.LoadAppConfigFile(fs, file)
 	if m["DB_USER"] != "dbuser" || m["EMAIL_PROVIDER"] != "smtp" {
 		t.Fatalf("unexpected map: %#v", m)
 	}
@@ -22,7 +23,7 @@ func TestLoadAppConfigFile(t *testing.T) {
 func TestLoadAppConfigFileMissing(t *testing.T) {
 	fs := core.UseMemFS(t)
 	file := "none.conf"
-	m := LoadAppConfigFile(fs, file)
+	m := config.LoadAppConfigFile(fs, file)
 	if len(m) != 0 {
 		t.Fatalf("expected empty map, got %#v", m)
 	}
