+++ handlers/admin/vars.go
@@ -3,7 +3,7 @@ package admin
 import (
 	"database/sql"
 
-	"github.com/arran4/goa4web/core"
+	"github.com/arran4/goa4web/config"
 	"github.com/arran4/goa4web/pkg/server"
 )
 
@@ -18,4 +18,4 @@ var DBPool *sql.DB
 
 // UpdateConfigKeyFunc is used to persist configuration changes. It should be
 // set by the main application on startup.
-var UpdateConfigKeyFunc func(fs core.FileSystem, path, key, value string) error
+var UpdateConfigKeyFunc func(fs config.FileSystem, path, key, value string) error
